requirements.txt
ansible==2.2.0
defusedxml==0.4.1
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.9.7
django-autoslug==1.9.3
django-crispy-forms==1.6.0
django-localflavor==1.3
django-paypal==0.3.2
django-registration-redux==1.4
django-toolbelt==0.0.1
gunicorn==19.6.0
jinja2==2.8
lxml==2.3
markupsafe==0.23
paramiko==2.0.1
Pillow==3.2.0
psycopg2==2.6.1
pyasn1==0.1.9
pycrypto==2.6.1
pytz==2016.4
pyyaml==3.11
requests==2.10.0
six==1.10.0
static3==0.7.0
stripe==1.35.0
virtualenv==13.1.0
whitenoise==3.2
